[iOS] Added plugin hook for iOS
As we need to embed pre-built frameworks GCDWebServer and XWalkView
into the product (while definitions in Cordova plugin.xml just link
them but won't embed them within), we will need to add this ability
into the Xcode project file when install this plugin and remove it
when uninstall.

And as the npm module 'xcode' hasn't provided this functionality to
embed frameworks into a specific target (the latest version of xcode
for now is v0.8.2), I have to implement it by hand.

The xcodeBuild.xcconfig is introduced to inject build settings into
the project. We set the iOS deployment target version to above 8.1
because XWalkView@1.2.0 only support iOS version above 8.1.

BUG=XWALK-4911